% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments_boxplot.R
\name{payments_boxplot}
\alias{payments_boxplot}
\title{Make a Boxplot of Payments by DRG Codes}
\usage{
payments_boxplot(x)
}
\arguments{
\item{x}{one of \code{"Average.Covered.Charges"}, \code{"Average.Total.Payments"} and
\code{"Average.Medicare.Payments"}}
}
\value{
An interactive boxplot
}
\description{
This function makes an interactive boxplot of payments data by DRG codes based on \code{DRG_data}.
}
\examples{
payments_boxplot("Average.Medicare.Payments")
}
