% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments_statistics.R
\name{payments_statistics}
\alias{payments_statistics}
\title{Calculates Statistics over DRG codes for Payments}
\usage{
payments_statistics(x = "Average.Medicare.Payments", op)
}
\arguments{
\item{x}{one of \code{"Average.Covered.Charges"}, \code{"Average.Total.Payments"} and
\code{"Average.Medicare.Payments"} (default value)}

\item{op}{one of statistics \code{mean}, \code{median}, and \code{sd}}
}
\value{
A tibble
}
\description{
This function calculates several statistics for payments data by DRG codes based on \code{DRG_data}.
}
\examples{
payments_statistics(op="mean")
}
